# backend/src/swagger-complete.yaml
openapi: 3.0.0
info:
  title: Jarvis Meeting Assistant API
  description: Complete API for meeting transcription, contact management, calendar sync, and automation
  version: 1.0.0
  contact:
    name: Jarvis Support
    email: support@jarvis.com
servers:
  - url: http://localhost:8080/api
    description: Local development server
  - url: https://your-production-url.com/api
    description: Production server
security:
  - UserAuth: []
components:
  securitySchemes:
    UserAuth:
      type: apiKey
      in: header
      name: x-user-id
      description: User ID for authentication
  schemas:
    Organization:
      type: object
      properties:
        id: { type: string }
        userId: { type: string }
        name: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    Workspace:
      type: object
      properties:
        id: { type: string }
        userId: { type: string }
        organizationId: { type: string }
        name: { type: string }
        description: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    Team:
      type: object
      properties:
        id: { type: string }
        userId: { type: string }
        workspaceId: { type: string }
        name: { type: string }
        description: { type: string }
        memberIds:
          type: array
          items: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    PortfolioStatus:
      type: object
      properties:
        completionPercentage: { type: number, format: float }
        totalTasks: { type: integer }
        completedTasks: { type: integer }
        projects:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              name: { type: string }
              completionPercentage: { type: number, format: float }
    Portfolio:
      type: object
      properties:
        id: { type: string }
        userId: { type: string }
        workspaceId: { type: string }
        name: { type: string }
        description: { type: string }
        projectIds:
          type: array
          items: { type: string }
        status:
          $ref: '#/components/schemas/PortfolioStatus'
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    Section:
      type: object
      properties:
        id: { type: string }
        userId: { type: string }
        projectId: { type: string }
        name: { type: string }
        taskIds:
          type: array
          items: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
paths:
...
  /organizations:
    get:
      summary: List organizations
      tags: [Project Management]
      responses:
        '200':
          description: List of organizations
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Organization'
    post:
      summary: Create organization
      tags: [Project Management]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
      responses:
        '201':
          description: Organization created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  id: { type: string }
  /organizations/{id}:
    get:
      summary: Get organization
      tags: [Project Management]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Organization
          content:
            application/json:
              schema:
                type: object
                properties:
                  item:
                    $ref: '#/components/schemas/Organization'
        '404':
          description: Not found
    patch:
      summary: Update organization
      tags: [Project Management]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
      responses:
        '200':
          description: Updated
    delete:
      summary: Delete organization
      tags: [Project Management]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Deleted
  /workspaces:
    get:
      summary: List workspaces
      tags: [Project Management]
      parameters:
        - in: query
          name: organizationId
          schema: { type: string }
      responses:
        '200':
          description: List of workspaces
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Workspace'
    post:
      summary: Create workspace
      tags: [Project Management]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [organizationId, name]
              properties:
                organizationId: { type: string }
                name: { type: string }
                description: { type: string }
      responses:
        '201':
          description: Workspace created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  id: { type: string }
  /workspaces/{id}:
    get:
      summary: Get workspace
      tags: [Project Management]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Workspace
          content:
            application/json:
              schema:
                type: object
                properties:
                  item:
                    $ref: '#/components/schemas/Workspace'
        '404':
          description: Not found
    patch:
      summary: Update workspace
      tags: [Project Management]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
      responses:
        '200':
          description: Updated
    delete:
      summary: Delete workspace
      tags: [Project Management]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Deleted
  /teams:
    get:
      summary: List teams
      tags: [Project Management]
      parameters:
        - in: query
          name: workspaceId
          schema: { type: string }
      responses:
        '200':
          description: List of teams
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Team'
    post:
      summary: Create team
      tags: [Project Management]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [workspaceId, name]
              properties:
                workspaceId: { type: string }
                name: { type: string }
                description: { type: string }
                memberIds:
                  type: array
                  items: { type: string }
      responses:
        '201':
          description: Team created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  id: { type: string }
  /teams/{id}:
    get:
      summary: Get team
      tags: [Project Management]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Team
          content:
            application/json:
              schema:
                type: object
                properties:
                  item:
                    $ref: '#/components/schemas/Team'
        '404':
          description: Not found
    patch:
      summary: Update team
      tags: [Project Management]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
                memberIds:
                  type: array
                  items: { type: string }
      responses:
        '200':
          description: Updated
    delete:
      summary: Delete team
      tags: [Project Management]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Deleted
  /portfolios:
    get:
      summary: List portfolios
      tags: [Project Management]
      parameters:
        - in: query
          name: workspaceId
          schema: { type: string }
      responses:
        '200':
          description: List of portfolios
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Portfolio'
    post:
      summary: Create portfolio
      tags: [Project Management]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [workspaceId, name]
              properties:
                workspaceId: { type: string }
                name: { type: string }
                description: { type: string }
      responses:
        '201':
          description: Portfolio created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  id: { type: string }
  /portfolios/{id}:
    get:
      summary: Get portfolio
      tags: [Project Management]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Portfolio
          content:
            application/json:
              schema:
                type: object
                properties:
                  item:
                    $ref: '#/components/schemas/Portfolio'
        '404':
          description: Not found
    patch:
      summary: Update portfolio
      tags: [Project Management]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
                projectIds:
                  type: array
                  items: { type: string }
      responses:
        '200':
          description: Updated
    delete:
      summary: Delete portfolio
      tags: [Project Management]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Deleted
  /portfolios/{id}/rollup:
    post:
      summary: Calculate and store portfolio roll-up status
      tags: [Project Management]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Roll-up calculated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  status:
                    $ref: '#/components/schemas/PortfolioStatus'
  /projects:
    get:
      summary: List projects
      tags: [Project Management]
      parameters:
        - in: query
          name: portfolioId
          schema: { type: string }
        - in: query
          name: teamId
          schema: { type: string }
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'
    post:
      summary: Create project
      tags: [Project Management]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [portfolioId, name]
              properties:
                portfolioId: { type: string }
                teamId: { type: string }
                name: { type: string }
                description: { type: string }
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  id: { type: string }
  /projects/{id}:
    get:
      summary: Get project
      tags: [Project Management]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Project
          content:
            application/json:
              schema:
                type: object
                properties:
                  item:
                    $ref: '#/components/schemas/Project'
        '404':
          description: Not found
    patch:
      summary: Update project
      tags: [Project Management]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
                status:
                  type: string
                  enum: [not_started, in_progress, completed]
                completionPercentage: { type: number, format: float }
                sectionIds:
                  type: array
                  items: { type: string }
                teamId: { type: string }
      responses:
        '200':
          description: Updated
    delete:
      summary: Delete project
      tags: [Project Management]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Deleted
  /sections:
    get:
      summary: List sections
      tags: [Project Management]
      parameters:
        - in: query
          name: projectId
          schema: { type: string }
      responses:
        '200':
          description: List of sections
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Section'
    post:
      summary: Create section
      tags: [Project Management]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [projectId, name]
              properties:
                projectId: { type: string }
                name: { type: string }
      responses:
        '201':
          description: Section created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  id: { type: string }
  /sections/{id}:
    get:
      summary: Get section
      tags: [Project Management]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Section
          content:
            application/json:
              schema:
                type: object
                properties:
                  item:
                    $ref: '#/components/schemas/Section'
        '404':
          description: Not found
    patch:
      summary: Update section
      tags: [Project Management]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                taskIds:
                  type: array
                  items: { type: string }
      responses:
        '200':
          description: Updated
    delete:
      summary: Delete section
      tags: [Project Management]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Deleted
  /tasks:
    get:
      summary: List tasks
      tags: [Project Management]
      parameters:
        - in: query
          name: projectId
          schema: { type: string }
        - in: query
          name: sectionId
          schema: { type: string }
        - in: query
          name: assigneeId
          schema: { type: string }
        - in: query
          name: status
          schema:
            type: string
            enum: [not_started, in_progress, completed]
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
    post:
      summary: Create task
      tags: [Project Management]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sectionId, projectId, title]
              properties:
                sectionId: { type: string }
                projectId: { type: string }
                title: { type: string }
                description: { type: string }
                assigneeId: { type: string }
                tags:
                  type: array
                  items: { type: string }
                customFields:
                  type: object
                  additionalProperties: { type: string }
                subtasks:
                  type: array
                  items:
                    type: object
                    properties:
                      title: { type: string }
                      status:
                        type: string
                        enum: [not_started, in_progress, completed]
                dependencies:
                  type: array
                  items: { type: string }
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  id: { type: string }
...